Link info.plist only if it doesn't exist

+++ run-tests.sh
@@ -14,9 +14,14 @@ rm -rf $LOGPATH
 
 curdir=$(pwd)
 wdir="${curdir}/tests"
+info_linked=0
 
-# link info.plist to parent directory so `background.py` can find it
-ln -s "${wdir}/info.plist"
+
+if [[ ! -f "info.plist" ]]; then
+	# link info.plist to parent directory so `background.py` can find it
+	ln -s "${wdir}/info.plist"
+	info_linked=1
+fi
 
 cd "$wdir"
 
@@ -54,6 +59,8 @@ esac
 
 cd "$curdir"
 
-rm -f "info.plist"
+if [[ $info_linked -eq 1 ]]; then
+	rm -f "info.plist"
+fi
 
 exit $ret
